oop changes, taskmanager is the same as tasklist
