+++ data/echo.txt
@@ -1,6 +1,4 @@
 D | 0 | return book | 2019-12-02 1800
-T | 0 | sleep
 T | 0 | woop
 T | 0 | asidgh
-D | 0 | return book | 2019-12-02 1800
-D | 0 | emh | 2024-06-13 1550
+T | 0 | aisdhg
